+++ webapp/javascript/models/appNames.ts
@@ -5,7 +5,3 @@ import { modelToResult } from './utils';
 export const appNamesModel = z.array(z.string());
 
 export type AppNames = z.infer<typeof appNamesModel>;
-
-export function parse(a: unknown): Result<AppNames, ZodError> {
-  return modelToResult(appNamesModel, a);
-}