+++ webapp/javascript/ui/Sidebar.tsx
@@ -12,8 +12,6 @@ import {
   SubMenuProps,
 } from 'react-pro-sidebar';
 import { IconProps } from '@webapp/ui/Icon';
-
-import styles from './Sidebar.module.css';
 import './Sidebar.scss';
 
 export interface SidebarProps {
@@ -40,10 +38,7 @@ type Icon = React.ReactElement<IconProps>;
 
 export function MenuItem(props: MenuItemProps & { icon: Icon }) {
   const { icon } = props;
-  let { className } = props;
-  if (icon) {
-    className = `${className} ${styles.menuWithIcon}`;
-  }
+  const { className } = props;
 
   return <RProMenuItem {...props} className={className} />;
 }